Rename components/club-details to resource-details
