+++ client/components/resource-details/ResourceInfo.tsx
