+++ client/components/resource-details/index.ts
